Added Images (Final Commit)
Need to update README with embeded links, so technically one more
commit. RIP Power Up, 2018-2021

+++ src/main/java/frc/robot/subsystems/ChainLift.java
@@ -189,7 +189,7 @@ public class ChainLift extends SubsystemBase {
     }
   }
   protected int getEncoderPosition() {
-    return masterChainMotor.getSelectedSensorPosition();
+    return (int)masterChainMotor.getSelectedSensorPosition();
   }
 
   /*
